pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-git-versioning<2"]
build-backend = "setuptools.build_meta"

[project]
name = "uac"
dynamic = ["version"]
authors = [
    {name = "Great Scott Gadgets", email = "dev@greatscottgadgets.com"},
]
license = { text = "BSD" }
description = "USB Audio Class 2.0 Example for the Great Scott Gadgets Cynthion"
urls = { Source = "https://github.com/greatscottgadgets/cynthion-uac" }
readme = "README.md"
dependencies = [
    "cynthion~=0.2.2",
    "luna-usb~=0.2.1",
    "wave~=0.0.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
uac = "uac"
